New Generator.Data assembly for the delta Script and OutputSummary json formats
