+++ hotreload-utils.proj
@@ -1,6 +1,7 @@
 <Project Sdk="Microsoft.Build.Traversal/3.0.3">
   <ItemGroup>
     <ProjectReference Include="src/Microsoft.DotNet.HotReload.Utils.Generator/Microsoft.DotNet.HotReload.Utils.Generator.csproj" />
+    <ProjectReference Include="src/Microsoft.DotNet.HotReload.Utils.Generator.Data/Microsoft.DotNet.HotReload.Utils.Generator.Data.csproj" />
     <ProjectReference Include="src/Microsoft.DotNet.HotReload.Utils.Generator.Frontend/Microsoft.DotNet.HotReload.Utils.Generator.Frontend.csproj" />
     <ProjectReference Include="src/hotreload-delta-gen/src/hotreload-delta-gen.csproj" />
     <ProjectReference Include="src/Microsoft.DotNet.HotReload.Utils.Generator.BuildTool/Microsoft.DotNet.HotReload.Utils.Generator.BuildTool.csproj" />
