+++ src/Microsoft.DotNet.HotReload.Utils.Generator.Data/OutputSummary/OutputSummary.cs
